+++ tests/plasticity/stdp_worker_test.cpp
@@ -12,17 +12,19 @@ struct WorkerNet {
   spatial::Lattice lattice{2, 1, 1, 1.0, 7};
   spatial::NeighborIndex neighbors{lattice, 2.0f, 1};
   neural::NeuronPool pool{lattice, neural::kDefaultLIF, 1.0, 99};
-  synaptic::SynapseIndex syn_index{lattice, neighbors, pool};
+  std::shared_ptr<synaptic::SynapseIndex> syn_index =
+      std::make_shared<synaptic::SynapseIndex>(lattice, neighbors, pool);
+  std::atomic<std::shared_ptr<synaptic::SynapseIndex>> syn_store{syn_index};
   synaptic::SynapseParams syn_params{};
   STDPParams params{};
 };
 
 TEST(STDPWorkerTest, ProcessesEnqueuedSpikes) {
   WorkerNet net;
-  auto sid = net.syn_index.Outgoing(0)[0];
-  float w0 = net.syn_index.Get(sid).weight;
+  auto sid = net.syn_index->Outgoing(0)[0];
+  float w0 = net.syn_index->Get(sid).weight;
 
-  STDPWorker worker(net.syn_index, net.pool, net.syn_params, net.params);
+  STDPWorker worker(net.syn_store, net.pool, net.syn_params, net.params);
   worker.Start();
 
   // Pre then post; worker should process both and potentiate.
@@ -34,13 +36,13 @@ TEST(STDPWorkerTest, ProcessesEnqueuedSpikes) {
   std::this_thread::sleep_for(std::chrono::milliseconds(10));
   worker.Stop();
 
-  float w1 = net.syn_index.Get(sid).weight;
+  float w1 = net.syn_index->Get(sid).weight;
   EXPECT_GT(w1, w0);
 }
 
 TEST(STDPWorkerTest, StopsCleanly) {
   WorkerNet net;
-  STDPWorker worker(net.syn_index, net.pool, net.syn_params, net.params);
+  STDPWorker worker(net.syn_store, net.pool, net.syn_params, net.params);
   worker.Start();
   worker.Stop();  // should not hang
   SUCCEED();